+++ src/Model/Input/ActivityLogInput.php
@@ -2,7 +2,7 @@
 
 namespace Locastic\Loggastic\Model\Input;
 
-use Symfony\Component\Serializer\Annotation\Groups;
+use Symfony\Component\Serializer\Attribute\Groups;
 
 class ActivityLogInput implements ActivityLogInputInterface
 {